# [CONTP-535] optimise tagger server chunking function by processing chunks in place