--- a/comp/core/tagger/server/util.go
+++ b/comp/core/tagger/server/util.go
@@ -11,15 +11,44 @@ import (
 	pb "github.com/DataDog/datadog-agent/pkg/proto/pbgo/core"
 )
 
-// splitBySize splits the given slice into contiguous non-overlapping subslices such that
-// the size of each sub-slice is at most maxChunkSize.
-// The size of each item is calculated using computeSize
-//
-// This function assumes that the size of each single item of the initial slice is not larger than maxChunkSize
-func splitBySize[T any](slice []T, maxChunkSize int, computeSize func(T) int) [][]T {
+type sizeComputerFunc[T any] func(T) int
 
-	// TODO: return an iter.Seq[[]T] instead of [][]T once we upgrade to golang v1.23
-	// returning iter.Seq[[]T] has better performance in terms of memory consumption
+type consumeChunkFunc[T any] func(T) error
+
+// computeProtoEventSize returns the size of a tags stream event in bytes
+func computeTagsEventInBytes(event *pb.StreamTagsEvent) int { return proto.Size(event) }
+
+// processChunksInPlace splits the passed slice into contiguous chunks such that the total size of each chunk is at most maxChunkSize
+// and applies the consume function to each of these chunks
+//
+// The size of an item is computed with computeSize
+// If an item has a size large than maxChunkSize, it is placed in a singleton chunk (chunk with one item)
+//
+// The consume function is applied to different chunks in-place, without any need extra memory allocation
+func processChunksInPlace[T any](slice []T, maxChunkSize int, computeSize sizeComputerFunc[T], consume consumeChunkFunc[[]T]) error {
+	idx := 0
+	for idx < len(slice) {
+		chunkSize := computeSize(slice[idx])
+		j := idx + 1
+
+		for j < len(slice) {
+			eventSize := computeSize(slice[j])
+			if chunkSize+eventSize > maxChunkSize {
+				break
+			}
+			chunkSize += eventSize
+			j++
+		}
+
+		if err := consume(slice[idx:j]); err != nil {
+			return err
+		}
+		idx = j
+	}
+	return nil
+}
+
+func splitBySize[T any](slice []T, maxChunkSize int, computeSize func(T) int) [][]T {
 	var chunks [][]T
 	currentChunk := []T{}
 	currentSize := 0
@@ -40,11 +69,21 @@ func splitBySize[T any](slice []T, maxChunkSize int, computeSize func(T) int) []
 	return chunks
 }
 
-// splitEvents splits the array of events to chunks with at most maxChunkSize each
-func splitEvents(events []*pb.StreamTagsEvent, maxChunkSize int) [][]*pb.StreamTagsEvent {
-	return splitBySize(
-		events,
-		maxChunkSize,
-		func(event *pb.StreamTagsEvent) int { return proto.Size(event) },
-	)
+// processChunksWithSplit splits the passed slice into contiguous chunks such that the total size of each chunk is at most maxChunkSize
+// and then applies the consume function to each of these chunks
+//
+// The size of an item is computed with computeSize
+// If an item has a size large than maxChunkSize, it is placed in a singleton chunk (chunk with one item)
+//
+// Prefer using processChunksInPlace for better CPU and memory performance. This implementation is only kept for benchmarking purposes.
+func processChunksWithSplit[T any](slice []T, maxChunkSize int, computeSize sizeComputerFunc[T], consume consumeChunkFunc[[]T]) error {
+	chunks := splitBySize(slice, maxChunkSize, computeSize)
+
+	for _, chunk := range chunks {
+		if err := consume(chunk); err != nil {
+			return err
+		}
+	}
+
+	return nil
 }