--- a/comp/core/tagger/server/util_test.go
+++ b/comp/core/tagger/server/util_test.go
@@ -6,6 +6,7 @@
 package server
 
 import (
+	"reflect"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -16,18 +17,35 @@ type mockStreamTagsEvent struct {
 	size int
 }
 
-func TestSplitEvents(t *testing.T) {
+var computeSize = func(e mockStreamTagsEvent) int {
+	return e.size
+}
+
+func getConsumeFunc(slice *[][]int) consumeChunkFunc[[]mockStreamTagsEvent] {
+	return func(chunk []mockStreamTagsEvent) error {
+		ids := make([]int, 0, len(chunk))
+		for _, item := range chunk {
+			ids = append(ids, item.id)
+		}
+
+		*slice = append(*slice, ids)
+		return nil
+	}
+}
+
+func Test(t *testing.T) {
+
 	testCases := []struct {
 		name         string
 		events       []mockStreamTagsEvent
 		maxChunkSize int
-		expected     [][]mockStreamTagsEvent // Expecting indices of events in chunks for easier comparison
+		expected     [][]int // Expecting id's of events in chunks for easier comparison
 	}{
 		{
 			name:         "Empty input",
 			events:       []mockStreamTagsEvent{},
 			maxChunkSize: 100,
-			expected:     nil, // No chunks expected
+			expected:     [][]int{},
 		},
 		{
 			name: "Single event within chunk size",
@@ -35,9 +53,9 @@ func TestSplitEvents(t *testing.T) {
 				{id: 1, size: 50}, // Mock event with size 50
 			},
 			maxChunkSize: 100,
-			expected: [][]mockStreamTagsEvent{
+			expected: [][]int{
 				{
-					{id: 1, size: 50}, // One chunk with one event
+					1, // One chunk with one event
 				},
 			},
 		},
@@ -47,9 +65,9 @@ func TestSplitEvents(t *testing.T) {
 				{id: 1, size: 20}, {id: 2, size: 30}, {id: 3, size: 40}, // Total size = 90
 			},
 			maxChunkSize: 100,
-			expected: [][]mockStreamTagsEvent{
+			expected: [][]int{
 				{
-					{id: 1, size: 20}, {id: 2, size: 30}, {id: 3, size: 40}, // All events fit in one chunk
+					1, 2, 3, // All events fit in one chunk
 				},
 			},
 		},
@@ -59,13 +77,12 @@ func TestSplitEvents(t *testing.T) {
 				{id: 1, size: 40}, {id: 2, size: 50}, {id: 3, size: 60}, // Total size = 150
 			},
 			maxChunkSize: 100,
-			expected: [][]mockStreamTagsEvent{
+			expected: [][]int{
 				{
-					{id: 1, size: 40},
-					{id: 2, size: 50},
+					1, 2,
 				},
 				{
-					{id: 3, size: 60},
+					3,
 				}, // Last event in second chunk
 			},
 		},
@@ -75,8 +92,8 @@ func TestSplitEvents(t *testing.T) {
 				{id: 1, size: 50}, {id: 2, size: 50}, // Total size = 100
 			},
 			maxChunkSize: 100,
-			expected: [][]mockStreamTagsEvent{
-				{{id: 1, size: 50}, {id: 2, size: 50}}, // Both events fit exactly in one chunk
+			expected: [][]int{
+				{1, 2}, // Both events fit exactly in one chunk
 			},
 		},
 		{
@@ -85,21 +102,35 @@ func TestSplitEvents(t *testing.T) {
 				{id: 1, size: 100}, {id: 2, size: 101}, // One exactly fits, one exceeds
 			},
 			maxChunkSize: 100,
-			expected: [][]mockStreamTagsEvent{
-				{
-					{id: 1, size: 100},
-				},
-				{
-					{id: 2, size: 101},
-				},
+			expected: [][]int{
+				{1}, {2},
+			},
+		},
+		{
+			name: "Multiple items exceeding max chunk size",
+			events: []mockStreamTagsEvent{
+				{id: 1, size: 100}, {id: 2, size: 101}, {id: 3, size: 101},
+			},
+			maxChunkSize: 100,
+			expected: [][]int{
+				{1}, {2}, {3},
 			},
 		},
 	}
 
 	for _, testCase := range testCases {
 		t.Run(testCase.name, func(t *testing.T) {
-			chunks := splitBySize(testCase.events, testCase.maxChunkSize, func(e mockStreamTagsEvent) int { return e.size })
-			assert.Equal(t, testCase.expected, chunks)
+			slice := make([][]int, 0, len(testCase.expected))
+			processChunksInPlace(testCase.events, testCase.maxChunkSize, computeSize, getConsumeFunc(&slice))
+			if len(testCase.expected) > 0 || len(slice) > 0 {
+				assert.Truef(t, reflect.DeepEqual(testCase.expected, slice), "expected %v, found %v", testCase.expected, slice)
+			}
+
+			slice = make([][]int, 0, len(testCase.expected))
+			processChunksWithSplit(testCase.events, testCase.maxChunkSize, computeSize, getConsumeFunc(&slice))
+			if len(testCase.expected) > 0 || len(slice) > 0 {
+				assert.Truef(t, reflect.DeepEqual(testCase.expected, slice), "expected %v, found %v", testCase.expected, slice)
+			}
 		})
 	}
 }